+++ src/get-data/get-riotgames-api-data/config.py
@@ -1,0 +1,12 @@
+import os 
+
+
+PLAYER_CSV_PATH = os.path.join(
+    os.path.join(os.getcwd(), "dataset"),
+    "player.csv"
+    )
+
+MATCH_CSV_PATH = os.path.join(
+    os.path.join(os.getcwd(), "dataset"), 
+    "match.csv"
+)
