+++ src/get-data/get-riotgames-api-data/constant.py
@@ -54,3 +54,5 @@ API_TOKEN = "RGAPI-ac0be8be-ff93-44f0-a629-3acbc5089afc"
 HEADERS = {
     "X-Riot-Token" : API_TOKEN
 }
+
+
