+++ src/get-data/get-riotgames-api-data/dataset/player.csv
@@ -1,0 +1,1 @@
+gameCreation,gameDuration,gameEndTimestamp,gameId,gameMode,gameName,gameStartTimestamp,gameType,gameVersion,mapId,platformId,queueId,teams,tournamentCode,assists,baronKills,bountyLevel,12AssistStreakCount,abilityUses,acesBefore15Minutes,alliedJungleMonsterKills,baronTakedowns,blastConeOppositeOpponentCount,bountyGold,buffsStolen,completeSupportQuestInTime,controlWardsPlaced,damagePerMinute,damageTakenOnTeamPercentage,dancedWithRiftHerald,deathsByEnemyChamps,dodgeSkillShotsSmallWindow,doubleAces,dragonTakedowns,earliestDragonTakedown,earlyLaningPhaseGoldExpAdvantage,effectiveHealAndShielding,elderDragonKillsWithOpposingSoul,elderDragonMultikills,enemyChampionImmobilizations,enemyJungleMonsterKills,epicMonsterKillsNearEnemyJungler,epicMonsterKillsWithin30SecondsOfSpawn,epicMonsterSteals,epicMonsterStolenWithoutSmite,flawlessAces,fullTeamTakedown,gameLength,getTakedownsInAllLanesEarlyJungleAsLaner,goldPerMinute,hadAfkTeammate,hadOpenNexus,immobilizeAndKillWithAlly,initialBuffCount,initialCrabCount,jungleCsBefore10Minutes,junglerKillsEarlyJungle,junglerTakedownsNearDamagedEpicMonster,kTurretsDestroyedBeforePlatesFall,kda,killAfterHiddenWithAlly,killParticipation,killedChampTookFullTeamDamageSurvived,killsNearEnemyTurret,killsOnLanersEarlyJungleAsJungler,killsOnOtherLanesEarlyJungleAsLaner,killsOnRecentlyHealedByAramPack,killsUnderOwnTurret,killsWithHelpFromEpicMonster,knockEnemyIntoTeamAndKill,landSkillShotsEarlyGame,laneMinionsFirst10Minutes,laningPhaseGoldExpAdvantage,legendaryCount,lostAnInhibitor,maxCsAdvantageOnLaneOpponent,maxKillDeficit,maxLevelLeadLaneOpponent,moreEnemyJungleThanOpponent,multiKillOneSpell,multiTurretRiftHeraldCount,multikills,multikillsAfterAggressiveFlash,mythicItemUsed,outerTurretExecutesBefore10Minutes,outnumberedKills,outnumberedNexusKill,perfectDragonSoulsTaken,perfectGame,pickKillWithAlly,poroExplosions,quickCleanse,quickFirstTurret,quickSoloKills,riftHeraldTakedowns,saveAllyFromDeath,scuttleCrabKills,skillshotsDodged,skillshotsHit,snowballsHit,soloBaronKills,soloKills,soloTurretsLategame,stealthWardsPlaced,survivedSingleDigitHpCount,survivedThreeImmobilizesInFight,takedownOnFirstTurret,takedowns,takedownsAfterGainingLevelAdvantage,takedownsBeforeJungleMinionSpawn,takedownsFirst25Minutes,takedownsInAlcove,takedownsInEnemyFountain,teamBaronKills,teamDamagePercentage,teamElderDragonKills,teamRiftHeraldKills,teleportTakedowns,threeWardsOneSweeperCount,tookLargeDamageSurvived,turretPlatesTaken,turretTakedowns,turretsTakenWithRiftHerald,twentyMinionsIn3SecondsCount,unseenRecalls,visionScoreAdvantageLaneOpponent,visionScorePerMinute,wardTakedowns,wardTakedownsBefore20M,wardsGuarded,challenges,champExperience,champLevel,championId,championName,championTransform,consumablesPurchased,damageDealtToBuildings,damageDealtToObjectives,damageDealtToTurrets,damageSelfMitigated,deaths,detectorWardsPlaced,doubleKills,dragonKills,firstBloodAssist,firstBloodKill,firstTowerAssist,firstTowerKill,gameEndedInEarlySurrender,gameEndedInSurrender,goldEarned,goldSpent,individualPosition,inhibitorKills,inhibitorTakedowns,inhibitorsLost,item0,item1,item2,item3,item4,item5,item6,itemsPurchased,killingSprees,kills,lane,largestCriticalStrike,largestKillingSpree,largestMultiKill,longestTimeSpentLiving,magicDamageDealt,magicDamageDealtToChampions,magicDamageTaken,neutralMinionsKilled,nexusKills,nexusLost,nexusTakedowns,objectivesStolen,objectivesStolenAssists,participantId,pentaKills,physicalDamageDealt,physicalDamageDealtToChampions,physicalDamageTaken,profileIcon,puuid,quadraKills,riotIdName,riotIdTagline,role,sightWardsBoughtInGame,spell1Casts,spell2Casts,spell3Casts,spell4Casts,summoner1Casts,summoner1Id,summoner2Casts,summoner2Id,summonerId,summonerLevel,summonerName,teamEarlySurrendered,teamId,teamPosition,timeCCingOthers,timePlayed,totalDamageDealt,totalDamageDealtToChampions,totalDamageShieldedOnTeammates,totalDamageTaken,totalHeal,totalHealsOnTeammates,totalMinionsKilled,totalTimeCCDealt,totalTimeSpentDead,totalUnitsHealed,tripleKills,trueDamageDealt,trueDamageDealtToChampions,trueDamageTaken,turretKills,turretsLost,unrealKills,visionScore,visionWardsBoughtInGame,wardsKilled,wardsPlaced,win
